backend/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
python-dotenv
flask-jwt-extended
